+++ CHANGELOG.md
@@ -11,17 +11,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+- added SSO support in `sagemaker-studio` module
 - added VPC/subnets/sg config for multi-account project template to `sagemaker-templates-service-catalog` module
 - added `sagemaker-custom-kernel` module
 - added batch inference project template to `sagemaker-templates-service-catalog` module
 - added EFS removal policy to `mlflow-fargate` module
 - added `mwaa` module with example dag which demonstrates the MLOps in Airflow
+- added `sagemaker-hugging-face-endpoint` module
 
 ### **Changed**
 
+- add account/region props for project templates in `sagemaker-templates-service-catalog` module
+- fix `sagemaker-templates-service-catalog` model deploy role lookup issue & abalone_xgboost model registry permissions
+- update `sagemaker-custom-kernel` module IAM permissions
+- split `xgboost_abalone` and `model_deploy` project templates in `sagemaker-templates-service-catalog` module
+- add support for other AWS partitions
 - update MySQL instance to use T3 instance type
 - upgrade `cdk_ecr_deployment` version to fix the deprecated `go1.x` lambda runtime
-- add support for other AWS partitions
 
 ### **Removed**